Update the styling of the event detail screen
